requirements.txt
Django==1.7.1
Pillow==2.5.3
Unidecode==0.4.17
django-forms-bootstrap==3.0.0
django-uuslug==1.0.2
python-slugify==1.0.2
virtualenv==1.11.6
